+++ server/src/main/java/com/store/api/dto/StockAdjustmentRequest.java
@@ -6,6 +6,6 @@ import java.math.BigDecimal;
 @Data
 public class StockAdjustmentRequest {
     private Long itemId;
-    private Integer quantity; // positive = add, negative = remove
+    private BigDecimal quantity; // positive = add, negative = remove
     private String reason;
 }
